refactor(cli) update attributes
