+++ CLI/controllers/update.go
@@ -2,8 +2,73 @@ package controllers
 
 import (
 	"cli/models"
+	"cli/utils"
+	"fmt"
 	"net/http"
+	"strings"
 )
+
+func (controller Controller) UpdateObject(path, attr string, values []any) error {
+	var err error
+	switch attr {
+	case "areas":
+		_, err = controller.UpdateRoomAreas(path, values)
+	case "separators+":
+		_, err = controller.AddRoomSeparator(path, values)
+	case "pillars+":
+		_, err = controller.AddRoomPillar(path, values)
+	case "breakers+":
+		_, err = controller.AddRackBreaker(path, values)
+	case "pillars-", "separators-", "breakers-":
+		_, err = controller.DeleteInnerAttrObj(path, strings.TrimSuffix(attr, "-"), values[0].(string))
+	case "vlinks+", "vlinks-":
+		_, err = controller.UpdateVirtualLink(path, attr, values[0].(string))
+	case "domain", "tags+", "tags-":
+		isRecursive := len(values) > 1 && values[1] == "recursive"
+		_, err = controller.UpdateObj(path, map[string]any{attr: values[0]}, isRecursive)
+	case "tags", "separators", "pillars", "vlinks", "breakers":
+		err = fmt.Errorf(
+			"object's %[1]s can not be updated directly, please use %[1]s+= and %[1]s-=",
+			attr,
+		)
+	case "description":
+		_, err = controller.UpdateDescription(path, attr, values)
+	default:
+		_, err = controller.UpdateAttributes(path, attr, values)
+
+	}
+	return err
+}
+
+func (controller Controller) UpdateAttributes(path, attributeName string, values []any) (map[string]any, error) {
+	var attributes map[string]any
+	if attributeName == "slot" || attributeName == "content" {
+		vecStr := []string{}
+		for _, value := range values {
+			vecStr = append(vecStr, value.(string))
+		}
+		var err error
+		if vecStr, err = models.CheckExpandStrVector(vecStr); err != nil {
+			return nil, err
+		}
+		attributes = map[string]any{attributeName: vecStr}
+	} else {
+		if len(values) > 1 {
+			return nil, fmt.Errorf("attributes can only be assigned a single value")
+		}
+		if vconfigAttr, found := strings.CutPrefix(attributeName, VIRTUALCONFIG+"."); found {
+			if len(vconfigAttr) < 1 {
+				return nil, fmt.Errorf("invalid attribute name")
+			}
+			vAttr := map[string]any{vconfigAttr: values[0]}
+			attributes = map[string]any{VIRTUALCONFIG: vAttr}
+		} else {
+			attributes = map[string]any{attributeName: values[0]}
+		}
+	}
+
+	return controller.UpdateObj(path, map[string]any{"attributes": attributes}, false)
+}
 
 func (controller Controller) UpdateObj(pathStr string, data map[string]any, withRecursive bool) (map[string]any, error) {
 	obj, err := controller.GetObject(pathStr)
@@ -98,4 +163,169 @@ func (controller Controller) UpdateObj(pathStr string, data map[string]any, with
 	}
 
 	return resp.Body, nil
+}
+
+func (controller Controller) UpdateDescription(path string, attr string, values []any) (map[string]any, error) {
+	if len(values) != 1 {
+		return nil, fmt.Errorf("a single value is expected to update a description")
+	}
+	newDesc, err := utils.ValToString(values[0], "description")
+	if err != nil {
+		return nil, err
+	}
+	data := map[string]any{"description": newDesc}
+	return controller.UpdateObj(path, data, false)
+}
+
+func (controller Controller) UpdateVirtualLink(path string, attr string, value string) (map[string]any, error) {
+	if len(value) == 0 {
+		return nil, fmt.Errorf("an empty string is not valid")
+	}
+
+	obj, err := controller.GetObject(path)
+	if err != nil {
+		return nil, err
+	} else if obj["category"] != models.EntityToString(models.VIRTUALOBJ) {
+		return nil, fmt.Errorf("only virtual objects can have vlinks")
+	}
+
+	vlinks, hasVlinks := obj["attributes"].(map[string]any)["vlinks"].([]any)
+	if attr == "vlinks+" {
+		if !hasVlinks {
+			vlinks = []any{value}
+		} else {
+			vlinks = append(vlinks, value)
+		}
+	} else if attr == "vlinks-" {
+		if !hasVlinks {
+			return nil, fmt.Errorf("no vlinks defined for this object")
+		}
+		vlinks, err = removeVirtualLink(vlinks, value)
+		if err != nil {
+			return nil, err
+		}
+	} else {
+		return nil, fmt.Errorf("invalid vlink update command")
+	}
+
+	data := map[string]any{"vlinks": vlinks}
+	return controller.UpdateObj(path, map[string]any{"attributes": data}, false)
+}
+
+func removeVirtualLink(vlinks []any, vlinkToRemove string) ([]any, error) {
+	for i, vlink := range vlinks {
+		if vlink == vlinkToRemove {
+			vlinks = append(vlinks[:i], vlinks[i+1:]...)
+			return vlinks, nil
+		}
+	}
+	return nil, fmt.Errorf("vlink to remove not found")
+}
+
+// attribute must be "separators", "pillars" or "breakers"
+func (controller Controller) DeleteInnerAttrObj(path, attribute, name string) (map[string]any, error) {
+	obj, err := controller.GetObject(path)
+	if err != nil {
+		return nil, err
+	}
+	attributes := obj["attributes"].(map[string]any)
+	attrMap, ok := attributes[attribute].(map[string]any)
+	if !ok || attrMap[name] == nil {
+		return nil, fmt.Errorf("%s %s does not exist", attribute, name)
+	}
+	delete(attrMap, name)
+	attributes[attribute] = attrMap
+	fmt.Println(attributes)
+	return controller.UpdateObj(path, map[string]any{"attributes": attributes}, false)
+}
+
+func (controller Controller) AddRoomPillar(path string, values []any) (map[string]any, error) {
+	// check and create a pillar
+	name, newPillar, err := models.ValuesToPillar(values)
+	if err != nil {
+		return nil, err
+	}
+
+	// get room and update pillars
+	obj, err := controller.GetObject(path)
+	if err != nil {
+		return nil, err
+	}
+	attr := obj["attributes"].(map[string]any)
+	var keyExist bool
+	attr["pillars"], keyExist = AddToMap[models.Pillar](attr["pillars"], name, newPillar)
+	obj, err = controller.UpdateObj(path, map[string]any{"attributes": attr}, false)
+	if err != nil {
+		return nil, err
+	}
+	if keyExist {
+		fmt.Printf("Pillar %s replaced\n", name)
+	}
+	return obj, nil
+}
+
+func (controller Controller) AddRackBreaker(path string, values []any) (map[string]any, error) {
+	// create a breaker
+	name, newBreaker, err := models.ValuesToBreaker(values)
+	if err != nil {
+		return nil, err
+	}
+
+	// get rack and modify breakers
+	obj, err := controller.GetObject(path)
+	if err != nil {
+		return nil, err
+	}
+	attr := obj["attributes"].(map[string]any)
+	var keyExist bool
+	attr["breakers"], keyExist = AddToMap[models.Breaker](attr["breakers"], name, newBreaker)
+	obj, err = controller.UpdateObj(path, map[string]any{"attributes": attr}, false)
+	if err != nil {
+		return nil, err
+	}
+	if keyExist {
+		fmt.Printf("Breaker %s replaced\n", name)
+	}
+	return obj, nil
+}
+
+func (controller Controller) AddRoomSeparator(path string, values []any) (map[string]any, error) {
+	name, newSeparator, err := models.ValuesToSeparator(values)
+	if err != nil {
+		return nil, err
+	}
+	obj, err := controller.GetObject(path)
+	if err != nil {
+		return nil, err
+	}
+	attr := obj["attributes"].(map[string]any)
+	var keyExist bool
+	attr["separators"], keyExist = AddToMap[models.Separator](attr["separators"], name, newSeparator)
+	obj, err = controller.UpdateObj(path, map[string]any{"attributes": attr}, false)
+	if err != nil {
+		return nil, err
+	}
+	if keyExist {
+		fmt.Printf("Separator %s replaced\n", name)
+	}
+	return obj, nil
+}
+
+func (controller Controller) UpdateRoomAreas(path string, values []any) (map[string]any, error) {
+	if attributes, e := models.SetRoomAreas(values); e != nil {
+		return nil, e
+	} else {
+		return controller.UpdateObj(path, map[string]any{"attributes": attributes}, false)
+	}
+}
+
+// Helpers
+func AddToMap[T any](mapToAdd any, key string, val T) (map[string]any, bool) {
+	attrMap, ok := mapToAdd.(map[string]any)
+	if !ok {
+		attrMap = map[string]any{}
+	}
+	_, keyExist := attrMap[key]
+	attrMap[key] = val
+	return attrMap, keyExist
 }
