+++ CLI/parser/ast_test.go
@@ -10,7 +10,6 @@ import (
 	"time"
 
 	"github.com/stretchr/testify/assert"
-	"golang.org/x/exp/maps"
 )
 
 func TestValueNodeExecute(t *testing.T) {
@@ -352,175 +351,6 @@ func TestSelectObjectNodeExecuteReset(t *testing.T) {
 	assert.Nil(t, err)
 	assert.Nil(t, value)
 	assert.Len(t, controllers.State.ClipBoard, 0)
-}
-
-func TestSetRoomAreas(t *testing.T) {
-	_, mockAPI, _, _ := test_utils.SetMainEnvironmentMock(t)
-
-	room := test_utils.GetEntity("room", "room", "site.building", "domain")
-
-	roomResponse := test_utils.GetEntity("room", "room", "site.building", "domain")
-	test_utils.MockGetObject(mockAPI, room)
-
-	roomResponse["attributes"] = map[string]any{
-		"reserved":  []float64{1, 2, 3, 4},
-		"technical": []float64{1, 2, 3, 4},
-	}
-	test_utils.MockUpdateObject(mockAPI, map[string]interface{}{"attributes": map[string]interface{}{"reserved": []float64{1, 2, 3, 4}, "technical": []float64{1, 2, 3, 4}}}, roomResponse)
-
-	reservedArea := []float64{1, 2, 3, 4}
-	technicalArea := []float64{1, 2, 3, 4}
-	value, err := setRoomAreas("/Physical/site/building/room", []any{reservedArea, technicalArea})
-
-	assert.Nil(t, err)
-	assert.NotNil(t, value)
-}
-
-func TestSetLabel(t *testing.T) {
-	_, mockAPI, _, _ := test_utils.SetMainEnvironmentMock(t)
-
-	room := test_utils.GetEntity("rack", "rack", "site.building.room", "domain")
-	test_utils.MockGetObject(mockAPI, room)
-	value, err := setLabel("/Physical/site/building/room/rack", []any{"myLabel"}, false)
-
-	assert.Nil(t, err)
-	assert.Nil(t, value)
-}
-
-func TestAddToMap(t *testing.T) {
-	newMap, replaced := addToMap[int](map[string]any{"a": 3}, "b", 10)
-
-	assert.Equal(t, map[string]any{"a": 3, "b": 10}, newMap)
-	assert.False(t, replaced)
-
-	newMap, replaced = addToMap[int](newMap, "b", 15)
-	assert.Equal(t, map[string]any{"a": 3, "b": 15}, newMap)
-	assert.True(t, replaced)
-}
-
-func TestRemoveFromStringMap(t *testing.T) {
-	newMap, deleted := removeFromStringMap[int]("{\"a\":3,\"b\":10}", "b")
-
-	assert.Equal(t, "{\"a\":3}", newMap)
-	assert.True(t, deleted)
-
-	newMap, deleted = removeFromStringMap[int](newMap, "b")
-	assert.Equal(t, "{\"a\":3}", newMap)
-	assert.False(t, deleted)
-}
-
-func TestAddRoomSeparatorOrPillarError(t *testing.T) {
-	tests := []struct {
-		name         string
-		addFunction  func(string, []any) (map[string]any, error)
-		values       []any
-		errorMessage string
-	}{
-		{"AddRoomSeparator", addRoomSeparator, []any{"mySeparator"}, "4 values (name, startPos, endPos, type) expected to add a separator"},
-		{"AddRoomPillar", addRoomPillar, []any{"myPillar"}, "4 values (name, centerXY, sizeXY, rotation) expected to add a pillar"},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			obj, err := tt.addFunction("/Physical/site/building/room", tt.values)
-
-			assert.Nil(t, obj)
-			assert.NotNil(t, err)
-			assert.ErrorContains(t, err, tt.errorMessage)
-		})
-	}
-}
-
-func TestAddRoomSeparatorOrPillarWorks(t *testing.T) {
-	tests := []struct {
-		name          string
-		addFunction   func(string, []any) (map[string]any, error)
-		values        []any
-		newAttributes map[string]any
-	}{
-		{"AddRoomSeparator", addRoomSeparator, []any{"mySeparator", []float64{1., 2.}, []float64{1., 2.}, "wireframe"}, map[string]interface{}{"separators": map[string]interface{}{"mySeparator": Separator{StartPos: []float64{1, 2}, EndPos: []float64{1, 2}, Type: "wireframe"}}}},
-		{"AddRoomPillar", addRoomPillar, []any{"myPillar", []float64{1., 2.}, []float64{1., 2.}, 2.5}, map[string]interface{}{"pillars": map[string]interface{}{"myPillar": Pillar{CenterXY: []float64{1, 2}, SizeXY: []float64{1, 2}, Rotation: 2.5}}}},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			_, mockAPI, _, _ := test_utils.SetMainEnvironmentMock(t)
-			room := test_utils.GetEntity("room", "room", "site.building", "domain")
-
-			test_utils.MockGetObject(mockAPI, room)
-			test_utils.MockGetObject(mockAPI, room)
-
-			room["attributes"] = tt.newAttributes
-			test_utils.MockUpdateObject(mockAPI, map[string]interface{}{"attributes": tt.newAttributes}, room)
-
-			obj, err := tt.addFunction("/Physical/site/building/room", tt.values)
-			assert.NotNil(t, obj)
-			assert.Nil(t, err)
-		})
-	}
-}
-
-func TestDeleteRoomPillarOrSeparatorWithError(t *testing.T) {
-	tests := []struct {
-		name          string
-		attributeName string
-		separatorName string
-		errorMessage  string
-	}{
-		{"InvalidArgument", "other", "separator", "others separator does not exist"},
-		{"SeparatorDoesNotExist", "separator", "mySeparator", "separators mySeparator does not exist"},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			_, mockAPI, _, _ := test_utils.SetMainEnvironmentMock(t)
-
-			room := test_utils.GetEntity("room", "room", "site.building", "domain")
-			test_utils.MockGetObject(mockAPI, room)
-			obj, err := deleteInnerAttrObj("/Physical/site/building/room", tt.attributeName+"s", tt.separatorName)
-
-			assert.Nil(t, obj)
-			assert.NotNil(t, err)
-			assert.ErrorContains(t, err, tt.errorMessage)
-		})
-	}
-}
-
-func TestDeleteRoomPillarOrSeparatorSeparator(t *testing.T) {
-	_, mockAPI, _, _ := test_utils.SetMainEnvironmentMock(t)
-
-	room := test_utils.GetEntity("room", "room", "site.building", "domain")
-	room["attributes"].(map[string]any)["separators"] = map[string]interface{}{"mySeparator": Separator{StartPos: []float64{1, 2}, EndPos: []float64{1, 2}, Type: "wireframe"}}
-
-	updatedRoom := test_utils.GetEntity("room", "room", "site.building", "domain")
-	updatedRoom["attributes"] = map[string]any{"separators": map[string]interface{}{}}
-
-	test_utils.MockGetObject(mockAPI, room)
-	test_utils.MockGetObject(mockAPI, room)
-
-	test_utils.MockUpdateObject(mockAPI, map[string]interface{}{"attributes": map[string]interface{}{"separators": map[string]interface{}{}}}, updatedRoom)
-	obj, err := deleteInnerAttrObj("/Physical/site/building/room", "separators", "mySeparator")
-
-	assert.Nil(t, err)
-	assert.NotNil(t, obj)
-}
-
-func TestDeleteRoomPillarOrSeparatorPillar(t *testing.T) {
-	_, mockAPI, _, _ := test_utils.SetMainEnvironmentMock(t)
-
-	room := test_utils.GetEntity("room", "room", "site.building", "domain")
-	room["attributes"].(map[string]any)["pillars"] = map[string]interface{}{"myPillar": Pillar{CenterXY: []float64{1, 2}, SizeXY: []float64{1, 2}, Rotation: 2.5}}
-
-	updatedRoom := maps.Clone(room)
-	updatedRoom["attributes"] = map[string]any{"pillars": map[string]interface{}{}}
-
-	test_utils.MockGetObject(mockAPI, room)
-	test_utils.MockGetObject(mockAPI, room)
-	test_utils.MockUpdateObject(mockAPI, map[string]interface{}{"attributes": map[string]interface{}{"pillars": map[string]interface{}{}}}, updatedRoom)
-	obj, err := deleteInnerAttrObj("/Physical/site/building/room", "pillars", "myPillar")
-
-	assert.Nil(t, err)
-	assert.NotNil(t, obj)
 }
 
 func TestUpdateObjNodeExecuteUpdateDescription(t *testing.T) {
