+++ CLI/parser/astutil_test.go
@@ -151,14 +151,6 @@ func TestEvalNodeArr(t *testing.T) {
 	assert.ErrorContains(t, err, "Error unexpected element")
 }
 
-func TestErrorResponder(t *testing.T) {
-	err := errorResponder("reserved", "4", false)
-	assert.ErrorContains(t, err, "Invalid reserved attribute provided. It must be an array/list/vector with 4 elements. Please refer to the wiki or manual reference for more details on how to create objects using this syntax")
-
-	err = errorResponder("reserved", "4", true)
-	assert.ErrorContains(t, err, "Invalid reserved attributes provided. They must be arrays/lists/vectors with 4 elements. Please refer to the wiki or manual reference for more details on how to create objects using this syntax")
-}
-
 func TestFiltersToMapString(t *testing.T) {
 	filters := map[string]node{
 		"tag": &valueNode{"my-tag"},
